backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.1
python-dotenv==1.0.0
pinecone-client==2.2.4
cohere==4.37
openai==1.3.7
alembic==1.13.1
tenacity==8.2.3
langchain-text-splitters==0.0.1
requests==2.31.0
